NutriSense/backend/service-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.10.1
motor==3.6.0